fix: optimize fetch data detail page
