+++ src/components/CardDetail.tsx
@@ -4,7 +4,11 @@
  */
 /** @jsxImportSource @emotion/react */
 
+import { useEffect, useState } from "react";
 import { css } from "@emotion/react";
+import { useParams } from "react-router-dom";
+import { GET_ANIME_LIST_BY_ID } from "../utils/queryGraphApolloClient";
+import { useQuery } from "@apollo/client";
 
 const cardStyles = css`
   display: flex;
@@ -45,34 +49,45 @@ const imgStyles = css`
   border-bottom-left-radius: 8px;
 `;
 
-interface TCard {
-  data: {
-    [key: string]: any;
-  };
-}
+type DetailHomePageParams = {
+  id: string;
+};
 
-const CardDetail = ({ data }: TCard) => {
+const CardDetail = () => {
+  const { id } = useParams<DetailHomePageParams>();
+  const [dataCardsDetail, setDataCardsDetail] = useState<{
+    [key: string]: any;
+  }>([]);
+
+  const { data } = useQuery(GET_ANIME_LIST_BY_ID, {
+    variables: {
+      id: id,
+    },
+  });
+
+  useEffect(() => {
+    if (data !== undefined) {
+      setDataCardsDetail([data.Media]);
+    }
+  }, [data]);
+
   return (
     <>
-      {data.length !== 0 ? (
-        data.map((card: any, index: any) => (
-          <div css={cardStyles} key={index}>
-            <div css={leftColumnStyles}>
-              <img
-                css={imgStyles}
-                src={card.coverImage.large}
-                alt={card.title.romaji}
-              />
-            </div>
-            <div css={rightColumnStyles}>
-              <h2>{card.title.romaji}</h2>
-              <p>{card.description}</p>
-            </div>
+      {dataCardsDetail.map((card: any, index: any) => (
+        <div css={cardStyles} key={index}>
+          <div css={leftColumnStyles}>
+            <img
+              css={imgStyles}
+              src={card.coverImage.large}
+              alt={card.title.romaji}
+            />
           </div>
-        ))
-      ) : (
-        <div>No content Found</div>
-      )}
+          <div css={rightColumnStyles}>
+            <h2>{card.title.romaji}</h2>
+            <p>{card.description}</p>
+          </div>
+        </div>
+      ))}
     </>
   );
 };
